rename classes, change icons from bottom nav
