+++ app/src/main/res/layout/fragment_help.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.ui.dashboard.DashboardFragment">
+    tools:context=".activities.ui.help.HelpFragment">
 
     <TextView
         android:id="@+id/text_dashboard"
