+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,13 +7,13 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:id="@+id/navigation_help"
+        android:icon="@android:drawable/ic_menu_help"
+        android:title="@string/title_help" />
 
     <item
         android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:icon="@drawable/baseline_person_24"
+        android:title="@string/title_profile" />
 
 </menu>
