+++ app/src/main/res/values/strings.xml
@@ -4,8 +4,8 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="title_activity_home">HomeActivity</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="title_help">Ayuda</string>
+    <string name="title_profile">Perfil</string>
     <string name="title_course_detail">Detalle del curso</string>
     <string name="title_register">Registración</string>
     <string name="title_login">Ingreso</string>
